+++ addon/components/ui-message.js
@@ -213,16 +213,6 @@ export default Ember.Component.extend({
   /**
     Message title.
 
-    @property title
-    @type String
-    @default null
-    @deprecated Use `caption`.
-  */
-  title: null,
-
-  /**
-    Message title.
-
     @property caption
     @type String
     @default null